moved a bunch of methods from test align to test utils; added a mapping from link type to which element contains global var in this link; added a list of all currently available for easy access (this also helped to have a much neater test of all the links being present); added a top link test for indirect links; deleted unused code
