+++ automates/text_reading/src/main/resources/application.conf
@@ -60,7 +60,7 @@ CommentEngine {
 
 apps {
 
-  projectDir = "/home/alexeeva/Repos/automates"
+  projectDir = "/Users/alexeeva/Repos/automates"
   grfnFile = "path/to/PETPT_GrFN.json"
   inputDirectory = "./input/SIR/text"
 
@@ -93,7 +93,7 @@ apps {
 
 alignment {
   alignerType = "pairwisew2v"
-  w2vPath = "/home/alexeeva/Repos/automates/automates/text_reading/vectors.txt" // a word embeddings file (e.g., GloVe) with this header: <len_of_vocab><space><embedding_size>, e.g., "6B 50" (no quotes); use of different sets of embeddings will require adjusting alignment similarity thresholds, e.g., commentTextAlignmentScoreThreshold
+  w2vPath = "vectors.txt" // a word embeddings file (e.g., GloVe) with this header: <len_of_vocab><space><embedding_size>, e.g., "6B 50" (no quotes); use of different sets of embeddings will require adjusting alignment similarity thresholds, e.g., commentTextAlignmentScoreThreshold
   relevantArgs = ["variable", "description"]
 }
