+++ automates/text_reading/src/main/scala/org/clulab/aske/automates/apps/ExtractAndAlign.scala
@@ -63,6 +63,19 @@ object ExtractAndAlign {
   val COMMENT_TO_GLOBAL_VAR = "comment_to_gvar"
   val GLOBAL_VAR_TO_SVO = "gvar_to_svo"
 
+  val allLinkTypes = Seq(
+    SRC_TO_COMMENT,
+    GLOBAL_VAR_TO_UNIT_VIA_IDENTIFIER,
+    GLOBAL_VAR_TO_UNIT_VIA_CONCEPT,
+    GLOBAL_VAR_TO_PARAM_SETTING_VIA_IDENTIFIER,
+    GLOBAL_VAR_TO_PARAM_SETTING_VIA_CONCEPT,
+    GLOBAL_VAR_TO_INT_PARAM_SETTING_VIA_IDENTIFIER,
+    GLOBAL_VAR_TO_INT_PARAM_SETTING_VIA_CONCEPT,
+    EQN_TO_GLOBAL_VAR,
+    COMMENT_TO_GLOBAL_VAR
+//    GLOBAL_VAR_TO_SVO,
+  )
+
   val whereIsGlobalVar = Map[String, String](
     GLOBAL_VAR_TO_UNIT_VIA_IDENTIFIER -> "element_1",
     GLOBAL_VAR_TO_UNIT_VIA_CONCEPT -> "element_1",
@@ -73,6 +86,18 @@ object ExtractAndAlign {
     EQN_TO_GLOBAL_VAR -> "element_2",
     COMMENT_TO_GLOBAL_VAR -> "element_2",
     GLOBAL_VAR_TO_SVO -> "element_1"
+  )
+
+  val whereIsNotGlobalVar = Map[String, String](
+    GLOBAL_VAR_TO_UNIT_VIA_IDENTIFIER -> "element_2",
+    GLOBAL_VAR_TO_UNIT_VIA_CONCEPT -> "element_2",
+    GLOBAL_VAR_TO_PARAM_SETTING_VIA_IDENTIFIER -> "element_2",
+    GLOBAL_VAR_TO_PARAM_SETTING_VIA_CONCEPT -> "element_2",
+    GLOBAL_VAR_TO_INT_PARAM_SETTING_VIA_IDENTIFIER -> "element_2",
+    GLOBAL_VAR_TO_INT_PARAM_SETTING_VIA_CONCEPT -> "element_2",
+    EQN_TO_GLOBAL_VAR -> "element_1",
+    COMMENT_TO_GLOBAL_VAR -> "element_1",
+    GLOBAL_VAR_TO_SVO -> "element_2"
   )
 
   // labels and argument names
